# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Tailor CRM API
  version: 3.0.0
  description: |
    # Production-Ready REST API for Tailor Management SaaS

    **Last Updated:** May 10, 2026
    **Status:** Fully Implemented & Tested
    **Total Endpoints:** 80 (73 + 7 system)

    ## Response Format (ALL endpoints)

    Success:   { success: true,  message: "...", data: {} }
    Error:     { success: false, message: "...", error: { code: "...", details: {} } }
    Paginated: { success: true,  message: "...", data: { items: [], pagination: { page, limit, total } } }


    ## Field Naming Convention
    - Requests:  snake_case  (customer_id, delivery_date)
    - Responses: camelCase   (customerId, deliveryDate)

    ## Order Status Flow
    pending → in_progress → ready → delivered  (cannot skip)

    ## Error Codes
    INVALID_INPUT | UNAUTHORIZED | FORBIDDEN | NOT_FOUND
    DUPLICATE_PHONE | INVALID_STATUS_TRANSITION
    CUSTOMER_NOT_FOUND | ORDER_NOT_FOUND | SHOP_NOT_FOUND
    MEASUREMENT_NOT_FOUND | STAFF_NOT_FOUND | INVOICE_NOT_FOUND
    INTERNAL_ERROR | RATE_LIMIT_EXCEEDED | SERVICE_UNAVAILABLE

    ## Rate Limiting
    All requests are rate-limited per IP address:
    - **Global**: 100 requests / 15 minutes (default)
    - **Login**: 5 requests / 15 minutes
    - **OTP/Verify**: 3 requests / 15 minutes
    - **Refresh**: 10 requests / 15 minutes
    - **Upload**: 20 requests / 15 minutes
    
    Rate limit status in response headers:
    - `RateLimit-Limit`: Maximum requests allowed
...

...

...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        ## JWT Authentication with Token Rotation & Blacklist
        
        **Obtain tokens from:**
        - POST /auth/login (Firebase production)
        - POST /auth/login-test (Development only)
        
        **Token Structure:**
        - Access Token: 7-day expiry, contains userId, phone, shopId, sessionId, jti
        - Refresh Token: Long-lived, used for rotation
        
        **Token Rotation Behavior:**
        - Every refresh returns NEW access + refresh tokens
        - Old refresh token becomes invalid (single-use)
        - Prevents token reuse attacks
        - Session ID maintained across rotations
        
        **Security Features:**
        - Redis-backed session tracking
        - Access token blacklist on logout
        - Session activity validation
        - JTI (JWT ID) claim prevents replay attacks
        - Automatic revocation on suspicious activity
        
        **On 401 Unauthorized:**
        - Token expired? Call POST /auth/refresh-token
        - Still fails? Re-login with POST /auth/login
        - Token revoked? Session blacklisted, must re-login
        
        **Usage:**
        Authorization: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
security:
  - BearerAuth: []
...
paths:
...
  /subscription/verify:
    post:
      tags: [Subscription]
      summary: Verify subscription status — NO AUTH REQUIRED
      description: Check subscription status for a shop. Used by frontend on app load.
      operationId: verifySubscription
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shopId:
                  type: integer
            example:
              shopId: 1
      responses:
        '200':
          description: Subscription status verified
          content:
            application/json:
              example:
                success: true
                message: Subscription verified
                data:
                  shopId: 1
                  planType: "free"
                  status: "trial"
                  isActive: true
                  trialEndDate: "2026-05-29"
                  daysRemaining: 30
  /subscription/create:
    post:
      tags: [Subscription]
      summary: Create subscription
      operationId: createSubscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [planType]
              properties:
                planType:
                  type: string
                  enum: [free, basic, premium, enterprise]
                amount:
                  type: number
            example:
              planType: "free"
              amount: 0.00
      responses:
        '201':
          description: Subscription created
          content:
            application/json:
              example:
                success: true
                message: Subscription created successfully
                data:
                  id: 1
                  shopId: 1
                  planType: "free"
                  status: "trial"
                  startDate: "2026-04-29"
                  trialEndDate: "2026-05-29"
                  amount: 0.00
                  isActive: true
                  createdAt: "2026-04-29T10:00:00Z"
...
  /subscription/{subscriptionId}/status:
    put:
      tags: [Subscription]
      summary: Update subscription status
      operationId: updateSubscriptionStatus
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, inactive, cancelled]
            example:
              status: "active"
      responses:
        '200':
          description: Status updated
          content:
            application/json:
              example:
                success: true
                message: Subscription status updated
                data:
                  id: 1
                  status: "active"
                  updatedAt: "2026-04-29T10:00:00Z"
  /subscription/{subscriptionId}:
    delete:
      tags: [Subscription]
      summary: Cancel subscription
      operationId: cancelSubscription
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Subscription cancelled
          content:
            application/json:
              example:
                success: true
                message: Subscription cancelled successfully
                data:
                  id: 1
                  status: "cancelled"
                  cancelledAt: "2026-04-29T10:00:00Z"